requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Medical imaging
pydicom>=2.3.0
SimpleITK>=2.1.0
nibabel>=3.2.0
scikit-image>=0.18.0

# Machine learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep learning
torch>=1.12.0
torchvision>=0.13.0
tensorflow>=2.8.0
keras>=2.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
PIL>=8.3.0

# Jupyter and development
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Utilities
tqdm>=4.62.0
PyYAML>=6.0
click>=8.0.0
pathlib2>=2.3.0

# Medical image registration (optional, for advanced preprocessing)
ants>=0.1.0
nipy>=0.4.0

# Data augmentation
albumentations>=1.2.0
imgaug>=0.4.0
